requirements.txt
djangorestframework-serializer-extensions